+++ src/main/java/ru/nsu/usova/dipl/situation/db/repository/BatchRepository.java
@@ -5,5 +5,5 @@ import java.util.List;
 public interface BatchRepository<T> {
     Long getMaxId();
 
-    List<T> getAllByMinMaxId(Long minId, Long maxId);
+    List<T> getAllByMinMaxId(Integer minId, Long maxId);
 }
